requirements.txt
torch>=2.1
torchvision>=0.16
tqdm
numpy
scikit-learn
Pillow
